add tests for all payment endpoints
this commit wil also add a new table to the DB, namely memo
the motivation for this change is to differentiate between a message
stored on a lightning invoice and a personal description of the payment.
Memo is added to the payment request and is visible to both parties
of the lightning payment. Description is used for internal, personal
descriptions

add test for payments.GetByID

make payments.UpdateUserBalance exportable

improve formating in payments package

add tests for UpdateInvoiceStatus

add test for payments.GetAll