Add email verification to users table
In this commit we introduce a new field to the users table, which tells
us if the user has verified their email or not. In the auth middleware,
we fail the request if the user hasn't verified their email. We also add
a basic endpoint for verifying a users email, given a token. This
endpoint is called in the util functions in httptestutil for creating
and authenticating a user.