Add API endpoints for querying for TX fees
We add two new HTTP endpoints for querying for fees, one for blockchain
TXs and one for LN TXs. These endpoints are a bit hard to test properly,
so that's skipped for now.

For onchain TXs, we'd either need a testnet
node that's been running for a long time, as Bitcoin Core estimates fees
by looking at TXs in the mempool, and how they fare when new blocks come
around. Testnet sees very few TXs and irregular blocks, so you need the
node to run for a while. One solution (haven't tested this) might be to
take the `fee_estimates.dat` file from a running `bitcoind` dir, and
copy that over to a node running on regtest.

For LN TXs, we'd need to setup a test harness with more than two LN
nodes, as there's no point in querying for fees if there's a direct
connection.